+++ crates/ludtwig/src/check/rules.rs
@@ -9,7 +9,7 @@ use crate::check::rules::twig_block_name_snake_case::RuleTwigBlockNameSnakeCase;
 use crate::check::rules::twig_hash_key_no_quotes::RuleTwigHashKeyNoQuotes;
 use crate::check::rules::twig_logic_and::RuleTwigLogicAnd;
 use crate::check::rules::twig_logic_or::RuleTwigLogicOr;
-use crate::check::rules::twig_prefer_sw_extends_over_extends::RuleTwigShopwareExtends;
+use crate::check::rules::twig_prefer_shopware_extends::RuleTwigPreferShopwareExtends;
 use crate::check::rules::twig_string_quotation::RuleTwigStringQuotation;
 use crate::check::rules::twig_use_is_not_same_as::RuleTwigUseIsNotSameAs;
 use crate::check::rules::twig_use_is_same_as::RuleTwigUseIsSameAs;
@@ -30,7 +30,7 @@ mod twig_block_name_snake_case;
 mod twig_hash_key_no_quotes;
 mod twig_logic_and;
 mod twig_logic_or;
-mod twig_prefer_sw_extends_over_extends;
+mod twig_prefer_shopware_extends;
 mod twig_string_quotation;
 mod twig_use_is_not_same_as;
 mod twig_use_is_same_as;
@@ -52,7 +52,7 @@ pub static RULE_DEFINITIONS: &[&'static dyn Rule] = &[
     &RuleTwigStringQuotation,
     &RuleHtmlStringQuotation,
     &RuleTwigHashKeyNoQuotes,
-    &RuleTwigShopwareExtends,
+    &RuleTwigPreferShopwareExtends,
     &RuleTwigUseIsSameAs,
     &RuleTwigUseIsNotSameAs,
 ];