+++ crates/ludtwig/src/check/rules/twig_prefer_shopware_extends.rs
@@ -3,30 +3,34 @@ use ludtwig_parser::syntax::untyped::SyntaxNode;
 
 use crate::check::rule::{CheckResult, Rule, RuleExt, RuleRunContext, Severity};
 
-pub struct RuleTwigShopwareExtends;
+pub struct RuleTwigPreferShopwareExtends;
 
-impl Rule for RuleTwigShopwareExtends {
+impl Rule for RuleTwigPreferShopwareExtends {
     fn name(&self) -> &'static str {
-        "twig-prefer-sw-extends-over-extends"
+        "twig-prefer-shopware-extends"
     }
 
     fn check_node(&self, node: SyntaxNode, _ctx: &RuleRunContext) -> Option<Vec<CheckResult>> {
         let twig_extends = TwigExtends::cast(node)?;
+        let extends_keyword = twig_extends.get_extends_keyword()?;
 
         // not a valid twig operator
         let result = self
-            .create_result(Severity::Error, "prefer 'sw_extends' over 'extends'")
+            .create_result(
+                Severity::Warning,
+                "prefer 'sw_extends' over 'extends' in shopware",
+            )
             .primary_note(
                 twig_extends.syntax().text_range(),
-                "help: change 'extends' to 'sw_extends'",
+                "change this 'extends' to 'sw_extends'",
             )
             .suggestion(
-                twig_extends.syntax().text_range(),
+                extends_keyword.text_range(),
                 "sw_extends",
                 "Try this keyword instead",
             );
 
-        return Some(vec![result]);
+        Some(vec![result])
     }
 }
 
@@ -38,17 +42,17 @@ mod tests {
     #[test]
     fn rule_reports() {
         test_rule(
-            "twig-prefer-sw-extends-over-extends",
+            "twig-prefer-shopware-extends",
             "{% extends foo %}",
             expect![[r#"
-                error[twig-prefer-sw-extends-over-extends]: prefer 'sw_extends' over 'extends'
+                warning[twig-prefer-shopware-extends]: prefer 'sw_extends' over 'extends' in shopware
                   ┌─ ./debug-rule.html.twig:1:1
                   │
                 1 │ {% extends foo %}
                   │ ^^^^^^^^^^^^^^^^^
-                  │ │
-                  │ help: change 'extends' to 'sw_extends'
-                  │ Try this keyword instead: sw_extends
+                  │ │  │
+                  │ │  Try this keyword instead: sw_extends
+                  │ change this 'extends' to 'sw_extends'
 
             "#]],
         );
@@ -57,9 +61,9 @@ mod tests {
     #[test]
     fn rule_fixes() {
         test_rule_fix(
-            "twig-prefer-sw-extends-over-extends",
+            "twig-prefer-shopware-extends",
             "{% extends foo %}",
-            expect!["sw_extends"],
+            expect!["{% sw_extends foo %}"],
         );
     }
 }