+++ crates/ludtwig/CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT-VERSION
 - Added `twig-use-is-not-same-as` rule
 - Added `twig-use-is-same-as` rule
+- Added `twig-prefer-shopware-extends` rule
 
 # v0.7.0
 - Rewrite of almost the complete codebase to make ludtwig a full linter / formatter